feat(#233): rename classes in bytecode package according with the new naming strategy
